release 20241111

+++ polardbx-cdc-backup/src/main/java/com/aliyun/polardbx/binlog/backup/BinlogUploadManager.java
@@ -1,24 +1,19 @@
 /**
- * Copyright (c) 2013-2022, Alibaba Group Holding Limited;
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * </p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright (c) 2013-Present, Alibaba Group Holding Limited.
+ * All rights reserved.
+ *
+ * Licensed under the Server Side Public License v1 (SSPLv1).
  */
 package com.aliyun.polardbx.binlog.backup;
 
 import com.aliyun.polardbx.binlog.ConfigKeys;
 import com.aliyun.polardbx.binlog.DynamicApplicationConfig;
+import com.aliyun.polardbx.binlog.LabEventManager;
+import com.aliyun.polardbx.binlog.MetaDbDataSource;
 import com.aliyun.polardbx.binlog.SpringContextHolder;
 import com.aliyun.polardbx.binlog.dao.BinlogOssRecordDynamicSqlSupport;
 import com.aliyun.polardbx.binlog.dao.BinlogOssRecordMapper;
+import com.aliyun.polardbx.binlog.domain.TaskType;
 import com.aliyun.polardbx.binlog.domain.po.BinlogOssRecord;
 import com.aliyun.polardbx.binlog.enums.BinlogUploadStatus;
 import com.aliyun.polardbx.binlog.filesys.CdcFile;
@@ -33,14 +28,19 @@ import com.aliyun.polardbx.binlog.remote.io.IFileCursorProvider;
 import com.aliyun.polardbx.binlog.remote.io.IFileReader;
 import com.aliyun.polardbx.binlog.service.BinlogOssRecordService;
 import com.aliyun.polardbx.binlog.util.BinlogFileUtil;
-import com.aliyun.polardbx.binlog.util.CommonUtils;
+import com.aliyun.polardbx.binlog.util.LabEventType;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
+import lombok.Getter;
 import org.mybatis.dynamic.sql.SqlBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.MDC;
 
 import java.io.IOException;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.HashSet;
@@ -50,6 +50,8 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.LinkedBlockingQueue;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
@@ -68,6 +70,7 @@ import static com.aliyun.polardbx.binlog.SpringContextHolder.getObject;
  */
 public class BinlogUploadManager implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(BinlogUploadManager.class);
+    @Getter
     private final BinlogOssRecordMapper recordMapper;
     private final BinlogOssRecordService recordService;
     /**
@@ -78,10 +81,15 @@ public class BinlogUploadManager implements Runnable {
      * 本地产生的binlog文件的保存路径（不带group和stream）
      */
     private final String binlogRootPath;
+    private final TaskType taskType;
     /**
      * dumper的名称，用于单流场景下判断该dumper是否是leader
      */
     private final String taskName;
+    /**
+     * 拓扑版本号，用于单流场景下判断该dumper是否是leader
+     */
+    private final long version;
     /**
      * group name
      */
@@ -93,7 +101,7 @@ public class BinlogUploadManager implements Runnable {
     /**
      * cluster id
      */
-    private final String clusterId;
+    private String clusterId;
     /**
      * 用于保存各个流对应的LocalFileSystem
      */
@@ -118,17 +126,27 @@ public class BinlogUploadManager implements Runnable {
      * BinlogUploader的状态
      */
     private volatile boolean runnable = true;
+    private static final int LOCK_CONTENTION_ERROR_CODE = 1205;
+    private static final String KEEP_ALIVE_SQL = "select 1";
+    private static final int KEEP_ALIVE_INTERVAL = 10;
+    @Getter
+    private final ScheduledThreadPoolExecutor keepAliveExecutor;
+    @Getter
+    private final boolean isLabEnv;
 
     public BinlogUploadManager(StreamContext context, Map<String, MetricsObserver> metrics) {
-        this.recordMapper = SpringContextHolder.getObject(BinlogOssRecordMapper.class);
         this.recordService = SpringContextHolder.getObject(BinlogOssRecordService.class);
+        this.recordMapper = SpringContextHolder.getObject(BinlogOssRecordMapper.class);
         this.provider = SpringContextHolder.getObject(IFileCursorProvider.class);
         this.binlogRootPath = BinlogFileUtil.getRootPath(context.getTaskType(), context.getVersion());
+        this.taskType = context.getTaskType();
         this.taskName = context.getTaskName();
+        this.version = context.getVersion();
         this.uploadingFiles = new HashSet<>();
         this.group = context.getGroup();
         this.streamList = context.getStreamList();
         this.clusterId = DynamicApplicationConfig.getString(ConfigKeys.CLUSTER_ID);
+        this.isLabEnv = DynamicApplicationConfig.getBoolean(ConfigKeys.IS_LAB_ENV);
 
         this.metricsObserverMap = metrics;
 
@@ -151,6 +169,11 @@ public class BinlogUploadManager implements Runnable {
             new ThreadFactoryBuilder().setNameFormat("binlog-backup-upload-thread-%d").build(),
             new ThreadPoolExecutor.CallerRunsPolicy());
         this.uploadExecutor.allowCoreThreadTimeOut(true);
+
+        this.keepAliveExecutor = new ScheduledThreadPoolExecutor(maxThreadNum,
+            new ThreadFactoryBuilder().setNameFormat("binlog-backup-upload-keep-alive-thread-%d").build());
+        this.keepAliveExecutor.setKeepAliveTime(60L, TimeUnit.SECONDS);
+        this.keepAliveExecutor.allowCoreThreadTimeOut(true);
     }
 
     public void start() {
@@ -174,9 +197,10 @@ public class BinlogUploadManager implements Runnable {
                 try {
                     // 每隔一段时间进行一轮扫描上传
                     Thread.sleep(1000);
-                    if (noNeedToUpload()) {
+                    if (!RuntimeLeaderElector.isDumperMasterOrX(version, taskType, taskName)) {
                         continue;
                     }
+
                     dispatchUploadJobs(getFilesToUpload());
                 } catch (Throwable e) {
                     logger.error("binlog uploader meet an exception", e);
@@ -186,14 +210,6 @@ public class BinlogUploadManager implements Runnable {
         } finally {
             MDC.remove(MDC_THREAD_LOGGER_KEY);
         }
-    }
-
-    /**
-     * 单流场景，slave不需要上传binlog，因为有mater干
-     */
-    private boolean noNeedToUpload() {
-        return CommonUtils.isGlobalBinlog(group, streamList.get(0))
-            && !RuntimeLeaderElector.isDumperLeader(taskName);
     }
 
     private void alert(String fileName) {
@@ -241,7 +257,7 @@ public class BinlogUploadManager implements Runnable {
                     processUpload(record);
                     logger.info("binlog file " + record.getBinlogFile() + " is successfully uploaded.");
                 } catch (Throwable e) {
-                    logger.error("upload file failed!", e);
+                    logger.error("upload file: {} failed!", record.getBinlogFile(), e);
                     alert(record.getBinlogFile());
                 } finally {
                     // 如果上传中发生异常，将这个record从uploadingFiles中移除，下次扫描线程又能够扫到这个record没有上传成功
@@ -254,14 +270,57 @@ public class BinlogUploadManager implements Runnable {
         }
     }
 
-    private void processUpload(BinlogOssRecord record) throws IOException {
-        setUploadStatusToUploading(record);
-        deleteFileOnRemote(record);
-        doUpload(record);
-        setUploadStatusToSuccess(record);
+    /**
+     * processUpload 需要保证delete file 和 upload 的原子性
+     * 单流场景下，在reBalance的时候，可能会有多个dumper同时执行processUpload方法
+     * 需要避免线程1上传成功并更改状态为success后，线程2将oss上的文件删除这种情况发生
+     */
+    public void processUpload(BinlogOssRecord record) throws Exception {
+        if (clusterId == null) {
+            clusterId = DynamicApplicationConfig.getString(ConfigKeys.CLUSTER_ID);
+        }
+
+        ScheduledFuture<?> future = null;
+        try (Connection conn = getConnection()) {
+            try {
+                if (globalLock(record, conn)) {
+                    setUploadStatusToUploading(record);
+                    deleteFileOnRemote(record);
+
+                    // 由于上传过程会很久，需要每隔一定时间执行一次sql，避免连接被视为空闲销毁
+                    future =
+                        getKeepAliveExecutor().scheduleAtFixedRate(() -> keepAlive(conn, record), KEEP_ALIVE_INTERVAL,
+                            KEEP_ALIVE_INTERVAL, TimeUnit.SECONDS);
+
+                    doUpload(record);
+
+                    setUploadStatusToSuccess(record);
+                }
+            } catch (SQLException e) {
+                if (e.getErrorCode() == LOCK_CONTENTION_ERROR_CODE) {
+                    // 锁冲突，说明有其他线程正在上传，直接返回
+                    logger.warn("lock conflict, skip uploading file: " + record.getBinlogFile());
+                } else {
+                    throw e;
+                }
+            } finally {
+                // 结束线程
+                if (future != null) {
+                    future.cancel(true);
+                }
+                synchronized (conn) {
+                    globalUnLock(record, conn);
+                }
+            }
+        }
     }
 
-    private void doUpload(BinlogOssRecord record) throws IOException {
+    Connection getConnection() throws SQLException {
+        MetaDbDataSource metaDs = getObject("metaDataSource");
+        return DriverManager.getConnection(metaDs.getUrl(), metaDs.getUsername(), metaDs.getPassword());
+    }
+
+    public void doUpload(BinlogOssRecord record) throws IOException {
         IFileReader fetcher = new BinlogFileReader(record.getBinlogFile(),
             BinlogFileUtil.getFullPath(binlogRootPath, record.getGroupId(), record.getStreamId()),
             new BinlogFileStatusChecker(provider, record.getStreamId()));
@@ -272,22 +331,64 @@ public class BinlogUploadManager implements Runnable {
         binlogUploader.upload();
     }
 
-    private void deleteFileOnRemote(BinlogOssRecord record) {
+    public void deleteFileOnRemote(BinlogOssRecord record) {
         String remoteFileName = BinlogFileUtil.buildRemoteFilePartName(
             record.getBinlogFile(), record.getGroupId(), record.getStreamId());
         RemoteBinlogProxy.getInstance().deleteFile(remoteFileName);
     }
 
-    private void setUploadStatusToSuccess(BinlogOssRecord record) {
-        recordMapper.update(u -> u.set(BinlogOssRecordDynamicSqlSupport.uploadStatus)
+    public boolean globalLock(BinlogOssRecord record, Connection conn) throws SQLException {
+        ResultSet resultSet =
+            conn.createStatement().executeQuery("SELECT GET_LOCK('" + getLockName(record.getId()) + "',1)");
+        if (resultSet != null && resultSet.next()) {
+            return resultSet.getInt(1) == 1;
+        } else {
+            return false;
+        }
+    }
+
+    public void globalUnLock(BinlogOssRecord record, Connection conn) {
+        try {
+            conn.createStatement().executeQuery("SELECT RELEASE_LOCK('" + getLockName(record.getId()) + "')");
+        } catch (Exception e) {
+            logger.error("unlock upload file {} failed", record.getBinlogFile(), e);
+            if (isLabEnv) {
+                LabEventManager.logEvent(LabEventType.UPLOAD_UNLOCK_FAIL, record.getBinlogFile());
+            }
+        }
+    }
+
+    public String getLockName(int recordId) {
+        return "uploadFile" + Integer.toHexString(recordId);
+    }
+
+    public void setUploadStatusToSuccess(BinlogOssRecord record) {
+        getRecordMapper().update(u -> u.set(BinlogOssRecordDynamicSqlSupport.uploadStatus)
             .equalTo(BinlogUploadStatus.SUCCESS.getValue()).set(BinlogOssRecordDynamicSqlSupport.uploadHost)
             .equalTo(DynamicApplicationConfig.getString(ConfigKeys.INST_IP))
             .where(BinlogOssRecordDynamicSqlSupport.id, SqlBuilder.isEqualTo(record.getId())));
     }
 
-    private void setUploadStatusToUploading(BinlogOssRecord record) {
-        recordMapper.update(u -> u.set(BinlogOssRecordDynamicSqlSupport.uploadStatus)
+    public void setUploadStatusToUploading(BinlogOssRecord record) {
+        getRecordMapper().update(u -> u.set(BinlogOssRecordDynamicSqlSupport.uploadStatus)
             .equalTo(BinlogUploadStatus.UPLOADING.getValue())
             .where(BinlogOssRecordDynamicSqlSupport.id, SqlBuilder.isEqualTo(record.getId())));
+    }
+
+    private void keepAlive(Connection c, BinlogOssRecord record) {
+        try {
+            synchronized (c) {
+                if (!c.isClosed()) {
+                    if (isLabEnv) {
+                        logger.info("try to keep alive upload file {}.", record.getBinlogFile());
+                    }
+                    c.createStatement().execute(KEEP_ALIVE_SQL);
+                } else {
+                    logger.warn("try to keep alive on a closed connection, file: {}.", record.getBinlogFile());
+                }
+            }
+        } catch (SQLException e) {
+            logger.error("keep alive failed.", e);
+        }
     }
 }
